d/psvr: Fix conversion warnings

+++ src/xrt/drivers/psvr/psvr_device.c
@@ -1059,8 +1059,8 @@ psvr_device_create(struct hid_device_info *sensor_hid_info,
 	info.display.h_meters = 0.07f;
 	info.lens_horizontal_separation_meters = 0.13f / 2.0f;
 	info.lens_vertical_position_meters = 0.07f / 2.0f;
-	info.fov[0] = 85.0f * (M_PI / 180.0f);
-	info.fov[1] = 85.0f * (M_PI / 180.0f);
+	info.fov[0] = (float)(85.0 * (M_PI / 180.0));
+	info.fov[1] = (float)(85.0 * (M_PI / 180.0));
 
 	if (!u_device_setup_split_side_by_side(&psvr->base, &info)) {
 		PSVR_ERROR(psvr, "Failed to setup basic device info");
